ok de novo

+++ WebRestaurante/src/com/br/servlets/RemoverMesaServlet.java
@@ -7,7 +7,9 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import com.br.model.Cliente;
 import com.br.model.Mesa;
+import com.br.model.Usuario;
 import com.br.services.MesaService;
 
 //@WebServlet("/removerMesa")
@@ -19,6 +21,12 @@ public class RemoverMesaServlet extends HttpServlet {
 			response.sendRedirect("LoginSistema");
 			return;
 		}
+		Usuario usuario = (Usuario)request.getSession().getAttribute("usuario");
+		if(usuario instanceof Cliente){
+			response.sendRedirect("LoginSistema");
+			return;
+		}
+		
 		String id = request.getParameter("id");
 		MesaService.remover(new Mesa(new Long(id)));
 		response.sendRedirect("listarMesa");
